Revert "Add node information to SchemaRegistryProtocol (#1485)"
This reverts commit 10337c35f7b22b969ba921318f1c2002996a769d.
