+++ core/src/main/java/io/confluent/kafka/schemaregistry/storage/SchemaRegistryIdentity.java
@@ -17,11 +17,12 @@ package io.confluent.kafka.schemaregistry.storage;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.confluent.kafka.schemaregistry.rest.SchemaRegistryConfig;
-import io.confluent.kafka.schemaregistry.utils.JacksonMapper;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
+
+import io.confluent.kafka.schemaregistry.rest.SchemaRegistryConfig;
+import io.confluent.kafka.schemaregistry.utils.JacksonMapper;
 
 /**
  * The identity of a schema registry instance. The leader will store the json representation of its
@@ -160,13 +161,11 @@ public class SchemaRegistryIdentity {
   @Override
   public String toString() {
     StringBuilder sb = new StringBuilder();
-    sb.append("SchemaRegistryIdentity{");
     sb.append("version=" + this.version + ",");
     sb.append("host=" + this.host + ",");
     sb.append("port=" + this.port + ",");
     sb.append("scheme=" + this.scheme + ",");
     sb.append("leaderEligibility=" + this.leaderEligibility);
-    sb.append("}");
     return sb.toString();
   }
