+++ core/src/test/java/io/confluent/kafka/schemaregistry/LeaderElectorTest.java
@@ -14,14 +14,6 @@
  */
 package io.confluent.kafka.schemaregistry;
 
-import io.confluent.kafka.schemaregistry.client.rest.RestService;
-import io.confluent.kafka.schemaregistry.client.rest.entities.SchemaString;
-import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException;
-import io.confluent.kafka.schemaregistry.id.ZookeeperIdGenerator;
-import io.confluent.kafka.schemaregistry.rest.SchemaRegistryConfig;
-import io.confluent.kafka.schemaregistry.storage.Mode;
-import io.confluent.kafka.schemaregistry.storage.SchemaRegistryIdentity;
-import io.confluent.kafka.schemaregistry.utils.TestUtils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -37,6 +29,15 @@ import java.util.List;
 import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.Callable;
+
+import io.confluent.kafka.schemaregistry.client.rest.RestService;
+import io.confluent.kafka.schemaregistry.client.rest.entities.SchemaString;
+import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException;
+import io.confluent.kafka.schemaregistry.rest.SchemaRegistryConfig;
+import io.confluent.kafka.schemaregistry.storage.Mode;
+import io.confluent.kafka.schemaregistry.utils.TestUtils;
+import io.confluent.kafka.schemaregistry.id.ZookeeperIdGenerator;
+import io.confluent.kafka.schemaregistry.storage.SchemaRegistryIdentity;
 
 import static io.confluent.kafka.schemaregistry.CompatibilityLevel.FORWARD;
 import static io.confluent.kafka.schemaregistry.CompatibilityLevel.NONE;
@@ -345,12 +346,12 @@ public class LeaderElectorTest extends ClusterTestHarness {
                                    CompatibilityLevel.NONE.name, true, null);
       leaderApps.add(leader);
       leader.start();
-      TestUtils.waitUntilLeaderElectionCompletes(leaderApps);
+      waitUntilLeaderElectionCompletes(leaderApps);
     }
 
     // Kill the current leader and wait for reelection until no leaders are left
     while (leaderApps.size() > 0) {
-      RestApp reportedLeader = TestUtils.checkOneLeader(leaderApps);
+      RestApp reportedLeader = checkOneLeader(leaderApps);
       leaderApps.remove(reportedLeader);
 
       checkLeaderIdentity(followerApps, reportedLeader.myIdentity());
@@ -358,7 +359,7 @@ public class LeaderElectorTest extends ClusterTestHarness {
       checkNoneIsLeader(followerApps);
 
       reportedLeader.stop();
-      TestUtils.waitUntilLeaderElectionCompletes(leaderApps);
+      waitUntilLeaderElectionCompletes(leaderApps);
     }
 
     // All leaders are now dead
@@ -462,10 +463,10 @@ public class LeaderElectorTest extends ClusterTestHarness {
 
     // Stop everything in the leader cluster
     while (leaderApps.size() > 0) {
-      RestApp leader = TestUtils.findLeader(leaderApps);
+      RestApp leader = findLeader(leaderApps);
       leaderApps.remove(leader);
       leader.stop();
-      TestUtils.waitUntilLeaderElectionCompletes(leaderApps);
+      waitUntilLeaderElectionCompletes(leaderApps);
     }
 
     // Try to register a new schema - should fail
@@ -601,10 +602,10 @@ public class LeaderElectorTest extends ClusterTestHarness {
 
     // Stop everything in the leader cluster
     while (leaderApps.size() > 0) {
-      RestApp leader = TestUtils.findLeader(leaderApps);
+      RestApp leader = findLeader(leaderApps);
       leaderApps.remove(leader);
       leader.stop();
-      TestUtils.waitUntilLeaderElectionCompletes(leaderApps);
+      waitUntilLeaderElectionCompletes(leaderApps);
     }
 
     // Try to register a new schema - should fail
@@ -636,9 +637,20 @@ public class LeaderElectorTest extends ClusterTestHarness {
     }
   }
 
+  /** Return the first node which reports itself as leader, or null if none does. */
+  private static RestApp findLeader(Collection<RestApp> cluster) {
+    for (RestApp restApp: cluster) {
+      if (restApp.isLeader()) {
+        return restApp;
+      }
+    }
+
+    return null;
+  }
+
   /** Verify that no node in cluster reports itself as leader. */
   private static void checkNoneIsLeader(Collection<RestApp> cluster) {
-    assertNull("Expected none of the nodes in this cluster to report itself as leader.", TestUtils.findLeader(cluster));
+    assertNull("Expected none of the nodes in this cluster to report itself as leader.", findLeader(cluster));
   }
 
   /** Verify that all nodes agree on the expected leader identity. */
@@ -661,6 +673,57 @@ public class LeaderElectorTest extends ClusterTestHarness {
             expectedLeaderIdentity, leaderIdentity);
       }
     }
+  }
+
+  /**
+   * Return set of identities of all nodes reported as leader. Expect this to be a set of
+   * size 1 unless there is some pathological behavior.
+   */
+  private static Set<SchemaRegistryIdentity> getLeaderIdentities(Collection<RestApp> cluster) {
+    Set<SchemaRegistryIdentity> leaderIdentities = new HashSet<SchemaRegistryIdentity>();
+    for (RestApp app: cluster) {
+      if (app != null && app.leaderIdentity() != null) {
+        leaderIdentities.add(app.leaderIdentity());
+      }
+    }
+
+    return leaderIdentities;
+  }
+
+  /**
+   * Check that exactly one RestApp in the cluster reports itself as leader.
+   */
+  private static RestApp checkOneLeader(Collection<RestApp> cluster) {
+    int leaderCount = 0;
+    RestApp leader = null;
+    for (RestApp restApp: cluster) {
+      if (restApp.isLeader()) {
+        leaderCount++;
+        leader = restApp;
+      }
+    }
+
+    assertEquals("Expected one leader but found " + leaderCount, 1, leaderCount);
+    return leader;
+  }
+
+  private void waitUntilLeaderElectionCompletes(final Collection<RestApp> cluster) {
+    if (cluster == null || cluster.size() == 0) {
+      return;
+    }
+
+    Callable<Boolean> newLeaderElected = new Callable<Boolean>() {
+      @Override
+      public Boolean call() throws Exception {
+        boolean hasLeader = findLeader(cluster) != null;
+        // Check that new leader identity has propagated to all nodes
+        boolean oneReportedLeader = getLeaderIdentities(cluster).size() == 1;
+
+        return hasLeader && oneReportedLeader;
+      }
+    };
+    TestUtils.waitUntilTrue(
+        newLeaderElected, 15000, "A node should have been elected leader by now.");
   }
 
   private void waitUntilIdExists(final RestService restService, final int expectedId, String errorMsg) {
